Refactor Cart entity and related DTOs to change price type from Double to Integer; rename CreateCartCommand to CartCommand and implement editCart functionality in CartApplicationService and CartController.

+++ product-service/product-domain/product-application-service/src/main/java/com/fns/product/service/domain/dto/get/CartResponse.java
@@ -22,6 +22,6 @@ public class CartResponse {
     private final ProductSizes size;
     private final Product product;
     private final Integer quantity;
-    private final Double price;
+    private final Integer price;
 
 }
